+++ sme/src/main/java/sme/tech/innovators/sme/dto/response/OrderItemDto.java
@@ -3,6 +3,8 @@ package sme.tech.innovators.sme.dto.response;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class OrderItemDto {
@@ -12,7 +14,7 @@ public class OrderItemDto {
     private String title;
     private String sku;
     private Integer quantity;
-    private Integer unitPriceAmount;
-    private Integer totalAmount;
+    private BigDecimal unitPriceAmount;
+    private BigDecimal totalAmount;
     private String currency;
 }
